resilienteco_functions/requirements.txt
azure-functions
aiohttp
reportlab
